fisheye.hh re stabilized.

+++ tools.cc
@@ -92,7 +92,7 @@ int main(int argc, const char* argv[]) {
       outs[0] = outs[1] = outs[2] = b_mesh.template cast<float>();
       normalize<float>(outs, 1.);
       savep2or3<float>(argv[4], outs, false);
-      saveobj(points, delaunay, argv[5], sqrt(double(data[0].rows() * data[0].cols()) / double(3)));
+      saveobj(points, delaunay, argv[5], sqrt(double(data[0].rows() * data[0].cols()) / double(256)));
     }
     break;
   case 6:
